AT PATCH, DELETE
things are imperfect but roughly working.

currently if I add an annotation to a word that didn't have any, and then delete it, it causes an error.

if I add, then re-load from airtable, then delete, then it works.

in future would be good to clean things up in various ways.

also, currently hard-coded the speaker and audio source IDs, rather than dynamically looking them up. so no matter which speaker and audio source are selected, it will always try to send the same one. for now.